Move medias.py to tests/fixtures/
vtsearch/medias.py was test-only — it generates synthetic audio fixtures
and writes to the medias proxy dict for the test suite to consume. The
app.py import was vestigial (annotated F401, with a comment explaining it
only existed for app_module.init_medias() in conftest).

Move it to tests/fixtures/medias.py where it belongs. conftest now imports
init_medias directly and attaches it to app_module for the handful of
tests that still reach in via app_module.init_medias().
