+++ CLAUDE.md
@@ -92,7 +92,7 @@ If a failure is genuinely outside the scope of the current task (e.g. a flaky ne
 - `app.py` — Flask entry point, registers blueprints, startup logic, CLI argument parsing, per-request user context via `before_request` middleware, per-request dataset/model context resolution from `X-Dataset-Id`/`X-Detector-Id` headers
 - `vtsearch/auth/` — Authentication: `LoginProvider` ABC, `DefaultLoginProvider` (single-user, no-op), `get_current_user()`, `get_user_data_dir()`, `set_login_provider()`
 - `vtsearch/config.py` — Constants (CLAP_SAMPLE_RATE, paths, model IDs)
-- `vtsearch/medias.py` — Test media generation and embedding cache management
+- `tests/fixtures/medias.py` — Test media generation and embedding cache management (loaded by `tests/conftest.py`; not part of the runtime app)
 - `vtsearch/cli.py` — CLI utilities: autodetect (load dataset + detectors from settings, run inference, export results)
 - `vtsearch/settings.py` — Persistent settings (volume, inclusion, theme, enrich_descriptions, safe_thresholds, calibrate_count, calibration_fraction, audio_playing, swipe_animation, show_metadata, view_mode_left, view_mode_right, focus_mode_left, focus_mode_right, grid_icon_size_left, grid_icon_size_right, panel_pct_left, panel_pct_right, autoload_media_embedders, autorun_detectors, autopilot_enabled, hide_autopilot, autopilot_top_greens, autopilot_hard_reds, autopilot_resort_interval, autopilot_goal_diversity, saved_datasets_dir, detectors_dir, max_concurrent_dataset_downloads, max_concurrent_dataset_embeddings, settings_source); auto-saves to `data/settings.json`. When a `settings_source` is configured, every save also syncs to the source (with a `_syncing` guard to prevent circular re-export during import). At startup, `sync_from_settings_source()` auto-imports from the active source. Also contains `_apply_settings()` for applying a settings dict via `set_*` functions
 - `vtsearch/routes/` — Flask blueprints: `auth.py`, `eval.py`, `file_browser.py`, `labels.py`, `media_server.py`, `main.py`, `medias.py`, `sorting.py`, `processors.py` (extractor/localizer/pregen-processor CRUD + execution; sub-modules `processors_crud.py` and `processors_scoring.py`), `detector_scoring.py` (`/api/auto-detect`, `/api/find-label` against detector labelsets), `detector_find.py` (multi-dataset Find), `datasets.py` (with sub-module `datasets_ui.py`), `datasets_registry.py` (dataset registry CRUD at `/api/datasets/registry/*`), `exporters.py`, `label_importers.py`, `settings.py`, `settings_io.py`, `sync_sources.py`, `detectors.py` (on-disk labelset+query store at `/api/detectors/*`), `detectors_registry.py` (in-memory model registry at `/api/detectors/registry/*`, autorun toggle); shared utilities in `helpers.py`. Note: dataset-load orchestration lives at `vtsearch/datasets/load_pipeline.py` (background-task helpers, ConcurrencyGate, clip fix-up).
